Removed unnecessary steps when registering app
The steps had you copy the Client ID but no later step use it.  This causes confusion.  I pasted it into ida:Wtrealm instead of the App ID URI in the confusion and was confused when it did not work.  Based on the exception I found that if I prefixed the Client ID with `spn:` it would work, but that is not the right solution per these steps.